+++ account-session.js
@@ -22,6 +22,7 @@ var VALID_KEYS = [
   'inForgotPasswordFlow',
   'inChangePasswordFlow',
   'inMessageOnlyFlow',
+  'inCreateAccountFlow',
 
   'errorMessage',
   'infoMessage',
@@ -72,6 +73,7 @@ Accounts._loginButtonsSession = {
     this.set('inForgotPasswordFlow', false);
     this.set('inChangePasswordFlow', false);
     this.set('inMessageOnlyFlow', false);
+    this.set('inCreateAccountFlow', false);
     this.set('dropdownVisible', false);
     this.resetMessages();
   },
@@ -98,11 +100,6 @@ Accounts._loginButtonsSession = {
   // ensure that somethings displaying a message (info or error) is
   // visible.  if a dialog with messages is open, do nothing;
   // otherwise open the dropdown.
-  //
-  // notably this doesn't matter when only displaying a single login
-  // button since then we have an explicit message dialog
-  // (_loginButtonsMessageDialog), and dropdownVisible is ignored in
-  // this case.
   ensureMessageVisible: function () {
     if (!this.isMessageDialogVisible())
       this.set("dropdownVisible", true);