+++ accounts-vuetify.js
@@ -1,4 +1,4 @@
-import { meteormogulAccountLogin } from './account-login.js';
+import { LoginButtons } from './account-login.js';
 
 // Variables exported by this module can be imported by other packages and
 // applications. See accounts-vuetify-tests.js for an example of importing.