+++ CHANGELOG.md
@@ -1,14 +1,21 @@
 # Changelog for Meteor Mogul accounts-vuetify
 
-## [0.0.3] - 2018-01-27
+## [0.0.4] - 2018-02-28
+
+- Fixed dates in CHANGELOG.  It's February not January!
+- Login or Sign Up as a button.
+- Renamed components.
+- Refactored into account-components
+
+## [0.0.3] - 2018-02-27
 
 - Material Design icon fonts
 - Show and hide login form
 
-## [0.0.2] - 2018-01-27
+## [0.0.2] - 2018-02-27
 
 - Kinda sorta working.
 
-## [0.0.1] - 2018-01-27
+## [0.0.1] - 2018-02-27
 
 - initial layout